Add pagination to quote list
This commit introduces pagination to the quote list view. The Django
Paginator class is used to paginate the quotes. The number of items per
page is set to 4. The current page number is obtained from the GET
parameters. If the page number is not valid or leads to an EmptyPage
exception, it is reset to 1. The pagination data is added to the
svelte props and used in the Paginator and QuoteList svelte components.