Update quote list and actions
This commit updates the quote list view and actions. The URL for deleting
quotes has been corrected. The quote list now includes the 'id', 'quote',
'author__name', and 'created_by' fields. A check has been added to determine
if the current user is the owner of a quote. The success URL for the
QuoteDeleteView now checks for a 'next' parameter. The Paginator and QuoteList
components in the Svelte frontend have been updated to reflect these changes.