Refactor views and add authorizer check
This commit refactors the views in the Django application by splitting
them into two separate files: `authors.py` and `quotes.py`. This makes
the code more organized and easier to maintain.

Additionally, a new function `is_authorizer` is added to check if a user
belongs to the 'authorizer' group. This function is used in various
places in the code to control access to certain features.

The changes also include updates to the URL routing to reflect the
refactoring of the views.