Add nginx Dockerfile and related files
This commit adds a Dockerfile for the nginx proxy along with related
configuration files. The Dockerfile sets up an unprivileged nginx
container and copies the configuration files into it. It also sets up
a volume for static files. The `run.sh` script is used to substitute
environment variables into the nginx configuration at runtime.