Implement dynamic theme switching in Django templates
This commit introduces dynamic theme switching in Django templates. The
`get_theme` function was added to `util.py` which fetches the current theme
from cookies. This theme is then passed as a prop to the `DarkMode` Svelte
component. The `navbar.html` and `quotes_app.html` templates were updated
to use this new functionality. The `preferences.py` file was also modified
to remove an unnecessary `login_required` decorator.