Replace DarkMode with ThemeToggle in Django and Svelte
This commit replaces the DarkMode component with a new ThemeToggle
component in both Django templates and Svelte. The changes include
updating the `navbar.html` and `quotes_app.html` templates to use the
new ThemeToggle component. The `DarkMode.js` file was deleted and a new
`ThemeToggle.js` file was created in its place. The `DarkMode.svelte`
component was renamed to `ThemeToggle.svelte`. Lastly, the `vite.config.ts`
file was updated to reflect these changes.