Add language toggle and update translations
This commit introduces a new language toggle feature in the navbar,
allowing users to switch between English and Spanish. The translations
for various strings in the application have been updated to reflect
changes in the source code. A new admin script has also been added to
facilitate the creation and compilation of message files.