Add i18n support and update LanguageToggle
This commit adds internationalization (i18n) support to the Django app by
including the Django i18n URLs in the app's URL configuration.

In addition, the LanguageToggle component in the Svelte frontend has been
updated. The component now makes a POST request to the Django backend to
change the language. The 'next' parameter has been added to the request to
ensure correct redirection after the language change.

A new function, set_language2, has been added to the LanguageToggle component
for testing purposes. This function makes a GET request to change the language,
but it is not currently used in the application.